# 2W1-G6-FACTURACION/docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'tpi-scaffolding'","description":"'TPI Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:58288","description":"Generated server url"}],"paths":{"/api/v1/updatePaymentMethod":{"put":{"tags":["payment-controller"],"operationId":"Update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentMethod"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentMethod"}}}}}}},"/api/v1/invoice/delete/{id}":{"put":{"tags":["invoice-controller"],"operationId":"updateInvoiceStat","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InvoiceDTO"}}}}}}},"/api/v1/paymentMethods/{name}":{"post":{"tags":["payment-controller"],"operationId":"Create","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentMethod"}}}}}}},"/api/v1/invoice":{"post":{"tags":["invoice-controller"],"summary":"Register a new invoice on the system","description":"Return the invoice created with the creation date and ID and client CUIT associated","operationId":"CreateInvoice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInvoiceRequest"}}},"required":true},"responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation, invoice created","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Payments sum is minor than invoice totalAmount","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/reportsByClients/pdf":{"get":{"tags":["reports-controller"],"operationId":"generateReportByClientPdf","parameters":[{"name":"dateFrom","in":"query","required":true,"schema":{"type":"string"}},{"name":"dateTo","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}}},"/reportByPayement":{"get":{"tags":["reports-controller"],"operationId":"getListPaymentReport","parameters":[{"name":"dateFrom","in":"query","required":true,"schema":{"type":"string"}},{"name":"dateTo","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TypePaymentDto"}}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/api/v1/paymentMethods":{"get":{"tags":["payment-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaymentMethod"}}}}}}}},"/api/v1/paymentMethods/{id}":{"get":{"tags":["payment-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentMethod"}}}}}}},"/api/v1/invoice/{id}":{"get":{"tags":["invoice-controller"],"summary":"Find an invoice by id value","description":"Return the invoice by the id, if doesn't exists one returns 404","operationId":"getInvoiceById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/invoice/generate-pdf-bytes/{id}":{"get":{"tags":["invoice-controller"],"operationId":"generateInvoicePdfBytes","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}}},"/api/v1/invoice/details/{id}":{"get":{"tags":["invoice-controller"],"summary":"Find all the details related to and invoice, by invoice id value","description":"Return the Details by the by the invoiceId, if the invoice with id specified doesn't exists returns 404","operationId":"getInvoiceDetails","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/invoice/billed-months":{"get":{"tags":["reports-controller"],"operationId":"getMonthlyTotalAmount","parameters":[{"name":"dateFrom","in":"query","required":true,"schema":{"type":"string"}},{"name":"dateTo","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BilledMonthDto"}}}}}}}},"/api/v1/invoice/amount/by-date":{"get":{"tags":["invoice-controller"],"summary":"Find total amount billed in a date range","description":"search the total amount billed by two filters dateto and datefrom you can fill bought, use just one or search by them all  , if there isn't any invoice matching filters returns 404","operationId":"getBilledAmountByClientAndDate","parameters":[{"name":"dateFrom","in":"query","required":false,"schema":{"type":"string"}},{"name":"dateTo","in":"query","required":false,"schema":{"type":"string"}},{"name":"clientId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/invoice/all":{"get":{"tags":["invoice-controller"],"summary":"Find all the invoices","description":"Return the invoices registered on the system, if there isn't any invoice returns 404","operationId":"getAllInvoices","responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/invoice/all/filtered":{"get":{"tags":["invoice-controller"],"summary":"Find all the invoices according to the filters sent","description":"Return the invoices registered on the system just by the filters indicated, is not necessary to fill them all you can just apply some , if there isn't any invoice matching filters returns 404","operationId":"getInvoicesByFilter","parameters":[{"name":"dateFrom","in":"query","required":false,"schema":{"type":"string"}},{"name":"dateTo","in":"query","required":false,"schema":{"type":"string"}},{"name":"clientId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/ReportsByClients":{"get":{"tags":["reports-controller"],"operationId":"getById_1","parameters":[{"name":"dateFrom","in":"query","required":true,"schema":{"type":"string"}},{"name":"dateTo","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClientsReport"}}}}}}}},"/api/v1/paymentMethod/delete/{id}":{"delete":{"tags":["payment-controller"],"operationId":"deleteById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}}},"components":{"schemas":{"PaymentMethod":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"InvoiceDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"client":{"type":"string"},"clientId":{"type":"integer","format":"int64"},"status":{"type":"string"},"date":{"type":"string"},"type":{"type":"string"},"totalAmount":{"type":"number"},"clientType":{"type":"string"}}},"CreateInvoiceRequest":{"type":"object","properties":{"orderId":{"type":"integer","format":"int64"},"clientId":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["A","B","C","TICKET"]},"status":{"type":"string","enum":["PENDING","BILLED","CANCELED"]},"iva":{"type":"number"},"discountRequestList":{"type":"array","items":{"$ref":"#/components/schemas/DiscountRequest"}},"reservationId":{"type":"integer","format":"int64"},"paymentMethodList":{"type":"array","items":{"$ref":"#/components/schemas/PaymentsRequest"}},"details":{"type":"array","items":{"$ref":"#/components/schemas/DetailDTO"}}}},"DetailDTO":{"type":"object","properties":{"product":{"$ref":"#/components/schemas/ProductDTO"},"amount":{"type":"integer","format":"int32"},"measurementUnit":{"type":"string"}}},"DiscountRequest":{"type":"object","properties":{"monto":{"type":"number"},"description":{"type":"string"}}},"PaymentsRequest":{"type":"object","properties":{"amount":{"type":"number"},"paymentMethod":{"type":"integer","format":"int64"},"observations":{"type":"string"}}},"ProductDTO":{"type":"object","properties":{"product_id":{"type":"string"},"name":{"type":"string"},"price":{"type":"number"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"TypePaymentDto":{"type":"object","properties":{"formapago":{"type":"string"},"monto":{"type":"number"}}},"BilledMonthDto":{"type":"object","properties":{"month":{"type":"string"},"totalAmount":{"type":"number"}}},"ClientsReport":{"type":"object","properties":{"name":{"type":"string"},"lastName":{"type":"string"},"nro_doc":{"type":"integer","format":"int64"},"phone":{"type":"string"},"clientType":{"type":"string"},"totalAmount":{"type":"number"}}}}}}