# 2W1-G6-FACTURACION/src/main/java/ar/edu/utn/frc/tup/lciv/dtos/BilledMonthDto.java
package ar.edu.utn.frc.tup.lciv.dtos;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class BilledMonthDto {
    private String month;
    private BigDecimal totalAmount;
}