# 2W1-G6-FACTURACION/src/main/java/ar/edu/utn/frc/tup/lciv/models/PaymentMethod.java
package ar.edu.utn.frc.tup.lciv.models;

import jakarta.persistence.Column;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class PaymentMethod {
    private Long id;
    private String name;
}